feat: refactor TokenSelect component docs; update CryptoInput component style; CI error fix
