+++ packages/web3/src/token-select/index.md
@@ -17,14 +17,13 @@ Supports two types of token list transmission: `controlled` and `interface call`
 
 <code src="./demos/basic.tsx"></code>
 
-### Query Token List via Interface
+### Select with search
 
-<code src="./demos/dynamicTokenList.tsx"></code>
+<code src="./demos/withSearch.tsx"></code>
 
 ## API
 
 | 属性 | 描述 | 类型 | 默认值 | 版本 |
 | --- | --- | --- | --- | --- |
-| queryTokenList | Interface to query the token list, triggered each time the selector is opened | () => Promise<[Token](/components/types-cn#Token)[]> | - | - |
-| tokenList | Token list (controlled), has higher priority than `defaultTokenList` and `queryTokenList` | [Token](/components/types-cn#Token)[] | - | - |
+| tokenList | selectable token list | [Token](/components/types-cn#Token)[] | - | - |
 | ...props | Please Reference AntDesign `Select` 组件 | [Select Props](https://ant-design.antgroup.com/components/select-cn#select-props) | - | - |
